Little fix MinFinReciever and NbyReciever
