+++ RestLib/Recievers/MinFinUkraine/MinFinReciever.cs
@@ -3,7 +3,6 @@ using log4net;
 using Newtonsoft.Json;
 using RestLib.Recievers.MinFinUkraine.Models;
 using RestSharp;
-using RestSharp.Authenticators;
 using System;
 using System.Linq;
 using System.Reflection;
@@ -11,20 +10,18 @@ using System.Threading.Tasks;
 
 namespace RestLib.Recievers.MinFinUkraine
 {
-    public class MinFinReciever : BaseReciever
+    public class MinFinReciever : BaseJsonReciever
     {
-        internal string json;
-        internal string apiKey;
         internal SummaryBankCurrencies currencies;
         public MinFinReciever(ILog iLog, IConfigurationProvider configurationProvider) : base(iLog, configurationProvider)
         {
-            apiKey = configurationProvider.ApiKey;
+            _resource += "/" + configurationProvider.ApiKey + "/";
         }
         public override string USD
         {
             get
             {
-                _log.Error($"Try to get USD exchange rate");
+                _log.Info($"Try to get USD exchange rate");
                 return GetExchangeRate("usd");
             }
         }
@@ -32,36 +29,8 @@ namespace RestLib.Recievers.MinFinUkraine
         {
             get
             {
-                _log.Error($"Try to get EUR exchange rate");
+                _log.Info($"Try to get EUR exchange rate");
                 return GetExchangeRate("eur");
-            }
-        }
-
-        internal async Task GetExchangeRateListAsync()
-        {
-            try
-            {
-                var request = new RestRequest(_resource + "/" + apiKey + "/", Method.Get);
-                var response = await client.GetAsync(request);
-                if (response.StatusCode == System.Net.HttpStatusCode.OK)
-                {
-                    if (response.Content is null)
-                    {
-                        _log.Error($"The responce content is empty");
-                        throw new Exception($"The responce content is empty");
-                    }
-                    json = response.Content;
-                }
-                else
-                {
-                    _log.Error($"The responce was broken. Status code is {response.StatusCode}");
-                    throw new Exception($"The responce was broken. Status code is {response.StatusCode}");
-                }
-            }
-            catch (Exception e)
-            {
-                _log.Error($"The error was ocured with the message '{e.Message}'");
-                throw;
             }
         }
 
@@ -78,7 +47,7 @@ namespace RestLib.Recievers.MinFinUkraine
                     currencies = JsonConvert.DeserializeObject<SummaryBankCurrencies>(json);
                 }
                 var pi = currencies.GetType().GetProperties(BindingFlags.Public | BindingFlags.Instance)
-                    .First(_ => _.GetCustomAttribute<JsonPropertyAttribute>().PropertyName == currencyCode  );
+                    .First(_ => _.GetCustomAttribute<JsonPropertyAttribute>().PropertyName == currencyCode);
                 var value = (SummaryBankCurrency)pi?.GetValue(currencies, null);
                 if (value is null || string.IsNullOrEmpty(value.Ask))
                 {
